- zscores are now compared in compare hits in QueryScore (bug) -   dbMatches counter in Sequence.h is now size_t

+++ src/commons/Sequence.h
@@ -15,7 +15,7 @@
 
 typedef struct {
         float kmersPerPos;
-        int dbMatches;
+        size_t dbMatches;
 } statistics_t;
 
 class Sequence
